+++ src/app/(app)/layout.tsx
@@ -2,6 +2,8 @@ import {ReactNode} from "react";
 import './globals.scss'
 import {Inter} from 'next/font/google'
 import {Content, DarkMode, Footer, Header} from "@/components";
+import Provider from "@/app/(app)/provider";
+import ProgressBarProvider from "@/components/layout/ProgressBarProvider";
 
 const inter = Inter({subsets: ['latin']})
 
@@ -14,10 +16,14 @@ export default function RootLayout({children}: { children: ReactNode }) {
     return (
         <html lang="en">
         <body className={`${inter.className} min-h-screen flex flex-col dark:bg-black`}>
-        <Header/>
-        <Content>{children}</Content>
-        <DarkMode/>
-        <Footer/>
+        <Provider>
+            <ProgressBarProvider>
+                <Header/>
+                <Content>{children}</Content>
+                <DarkMode/>
+                <Footer/>
+            </ProgressBarProvider>
+        </Provider>
         </body>
         </html>
     )