+++ src/assets/images/index.ts
@@ -1,0 +1,2 @@
+export {default as DarkLogo} from "./bookstore_logo.png";
+export {default as WhiteLogo} from "./bookstore_logo_white.png";