+++ src/components/index.tsx
@@ -1,6 +1,7 @@
 export {default as Header} from "./layout/Header";
 export {default as Footer} from "./layout/Footer";
 export {default as Content} from "./layout/Content";
+export {default as Logo} from "./layout/Logo";
 export {default as DarkMode} from "./layout/DarkMode";
 
 export {default as Container} from "./common/Container";
@@ -8,4 +9,5 @@ export {default as ActiveLink} from "./common/ActiveLink";
 
 export {default as BookCarousel} from "./pages/BookCarousel";
 export {default as BookImage} from "./pages/BookImage";
+export {default as Books} from "./pages/Books";
 export {default as Book} from "./pages/Book";