+++ src/components/layout/Footer.tsx
@@ -4,7 +4,7 @@ import {Container} from "@/components";
 const Footer = () => {
     return (
         <footer className="mt-auto my-2">
-            <Container className="text-center">
+            <Container className="text-center mt-5">
                 <span className="text-red-600">❤</span>
             </Container>
         </footer>