+++ src/types/common.d.ts
@@ -1,3 +1,5 @@
 interface IBookResponse<T> {
     book: T;
 }
+
+type IPreferColorScheme = "light" | "dark";